-- Added some calls to fetch followers as well as add a follower
-- Added unit tests showing adding a follower works

-- Fixed a bug where the salt was not matching as well as a unit test showing it matches

+++ tests/redistests.py
@@ -25,12 +25,26 @@ class RedisTests(unittest.TestCase):
         self.assertTrue(self.dbh.username_exists(self.user.username))
         self.assertTrue(self.dbh.user_id_exists(self.user.id))
 
-    def get_user(self):
-        """
-        Test to see if a user can be retrieved from redis.
-        """
         stored_user = self.dbh.get_user_by_id(self.user.id)
-        self.assertEqual(self.user, stored_user)
+        self.assertEqual(self.user.email, stored_user.email)
+        self.assertEqual(self.user.username, stored_user.username)
+        self.assertEqual(str(self.user.id), str(stored_user.id))
+        self.assertEqual(self.user.salt, stored_user.salt)
+        self.assertEqual(self.user.name, stored_user.name)
+        self.assertEqual(self.user.followers, stored_user.followers)
+        self.assertEqual(self.user.following, stored_user.following)
+        self.assertEqual(self.user.messages, stored_user.messages)
+
+        passwd = UserHelper.hash_password(self.user.password, stored_user.salt)
+        self.assertEqual(passwd, stored_user.password)
+
+    def test_add_follower(self):
+        """
+        Test adding and getting a follower
+        """
+        self.dbh.add_follower(2, self.user.id)
+        followers = self.dbh.get_follower_ids(self.user.id)
+        self.assertTrue('2' in followers)
 
     def tearDown(self):
         """
@@ -39,7 +53,9 @@ class RedisTests(unittest.TestCase):
         @return:
         @rtype:
         """
-        self.dbh.delete_user(self.user.id)
+        if self.dbh.user_id_exists(self.user.id):
+            self.dbh.delete_user(self.user.id)
+
         self.assertFalse(self.dbh.user_id_exists(self.user.id))
         self.assertFalse(self.dbh.email_exists(self.user.email))
         self.assertFalse(self.dbh.username_exists(self.user.username))
